requirements.txt
# Core AI
httpx==0.27.0
chromadb==1.2.0
ollama==0.2.1
google-generativeai==0.5.4

# ML/Embeddings
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2
langdetect==1.0.9

# Data handling
pyyaml==6.0.1
pydantic==2.6.1

# Database
sqlalchemy==2.0.25

# Utilities
pyyaml==6.0.1
pydantic==2.6.1
python-dotenv==1.0.1
loguru==0.7.2

# Development
pytest==8.0.0
black==24.1.1

# NumPy compatibility
numpy==1.24.3